% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytesMidiNotes.R
\name{bytesMidiNotes}
\alias{bytesMidiNotes}
\title{Generate midit byte stream for notes DF}
\usage{
bytesMidiNotes(midiNotesDF)
}
\arguments{
\item{midiNotesDF}{processed notes DF with midi bytes in each row}
}
\description{
simply concatenates all bytes in data frame.
Assumes they are in order of delta time
}
