% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytesMidiNotes.R
\name{bytesMidiTrack}
\alias{bytesMidiTrack}
\title{Generate bytes for one midi Track}
\usage{
bytesMidiTrack(notesDF, trackname, instrument)
}
\arguments{
\item{notesDF}{data frame of pitch, velocity, duration and beat}

\item{trackname}{char string}

\item{instrument}{char string}
}
\description{
Return bytes from a notes data frame constructing the track header, notes, end of track messages
}
