% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_midi.R
\name{create_midi}
\alias{create_midi}
\title{Create MIDI sequences.}
\usage{
create_midi(notes_arg, tempo = 120)
}
\arguments{
\item{notes_arg}{The sequence to be turned into hexadecimal MIDI.  This is the output of the drum_machine function or a modified drum sequence.}
}
\value{
A hexadecimal MIDI sequence to be entered into the write_midi function.
}
\description{
Converts musical sequences into hexadecimal MIDI sequences.
}
\examples{
# Shifts notes in the first track (hh, or hi-hat) of drum sequence to "G#1" with .5 probability.

drum_machine(hh = 1:16, kick = seq(1, 16, by = 4), snare = c(5, 13)) \%>\%
create_midi()


}
