% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drum_machine.R
\name{drum_machine}
\alias{drum_machine}
\title{Create drum sequence.}
\usage{
drum_machine(..., note_length = NULL, rep = 1)
}
\arguments{
\item{hh}{(optional) Takes numbers from 1 to 16, which represent the hi-hat's positions in a 16-step sequence.}

\item{kick}{(optional) Takes numbers from 1 to 16, which represent the kick's positions in a 16-step sequence.}

\item{snare}{(optional) Takes numbers from 1 to 16, which represent the snare's positions in a 16-step sequence.}
}
\value{
A named list with notes in positions corresponding to the hits.  This list has 'seq' as its class attribute, which means that it can be fed into the create_midi function to be converted into a hexadecimal MIDI sequence.
}
\description{
Creates drum sequence as named  list.  Emulates a 16-step drum sequencer.
}
\examples{
drum_machine(hh = 1:16, kick = seq(1,16, by = 4),
snare = c(5, 13))


}
