% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drum_shifter.R
\name{drum_shifter}
\alias{drum_shifter}
\title{Randomly shift notes in a MIDI sequence}
\usage{
drum_shifter(seq_arg = NULL, shift = NULL, prob = 1,
  position = "all", track_apply = NULL, ...)
}
\arguments{
\item{seq_arg}{The sequence to be modified.  This should be the output of the drum_machine function.}

\item{shift}{The note to shift to.  This should be a capital letter for the note and a number for the octave (e.g "C3").}

\item{prob}{(default = 1) The probability a note in the sequence will be shifted}

\item{position}{(default = "all") The position of items in the sequence to be shifted}

\item{track_apply}{An integer or vector of integers indicating which tracks (hh, kick, snare) to apply the shift to.}
}
\value{
A modified MIDI sequence to be entered into the create_midi function.
}
\description{
Takes sequence produced by drum_machine function and shifts notes in this sequence.
}
\examples{
# Shifts notes in the first track (hh, or hi-hat) of drum sequence to "G#1" with .5 probability.

drum_machine(hh = 1:16, kick = seq(1, 16, by = 4), snare = c(5, 13)) \%>\%
drum_shifter(., shift = "G#1", prob = .5, track_apply = 1)


}
