% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromMIDtoID.par.R
\name{FromMIDtoID.par}
\alias{FromMIDtoID.par}
\title{Convert mode-parameterized inverted Dirichlet paramaeters to inverted Dirichlet Parameters}
\usage{
FromMIDtoID.par(theta, gamma)
}
\arguments{
\item{theta}{Numeric vector of mode parameters, all elements must be positive.}

\item{gamma}{Numeric scalar, the dispersion parameter, must be positive.}
}
\value{
A numeric vector of length \code{d}, the concentration parameters \code{alpha}.
}
\description{
Converts mode-based inverted Dirichlet parameters, theta and gamma, to the parameters of an inverted Dirichlet distribution.
}
\examples{
theta <- c(0.2, 0.3)
gamma <- 0.5
FromMIDtoID.par(theta, gamma)

}
