% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiByteEncodings.R
\name{midiHeader}
\alias{midiHeader}
\title{Generate midiHeader
\code{<Header Chunk> = <chunk type><length><format><ntrks><division>}
These are NOT coded as var length integers}
\usage{
midiHeader()
}
\description{
Generate midiHeader
\code{<Header Chunk> = <chunk type><length><format><ntrks><division>}
These are NOT coded as var length integers
}
