% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiByteEncodings.R
\name{midiTrkHeader}
\alias{midiTrkHeader}
\title{Generate Midi Track Header}
\usage{
midiTrkHeader(trackLength = 100)
}
\arguments{
\item{trackLength}{length of track in bytes
NB track length needs to be calculated once all the midi notes have been coded}
}
\description{
Generate Midi Track Header
}
