% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_modify.R
\name{random_modify}
\alias{random_modify}
\title{Randomly modify MIDI sequence}
\usage{
random_modify(seq_arg = NULL, modifier = NULL, prob = 1,
  position = "all", track_apply = NULL, cc_parm = NULL,
  cc_map = NULL)
}
\arguments{
\item{seq_arg}{The sequence to be modified.  This should be the output of the drum_machine function.}

\item{modifier}{("double", "flam", "roll", "cc") The modification to apply.}

\item{prob}{(default = 1) The probability a note in the sequence will receive the modification.}

\item{position}{(default = "all") The position of items in the sequence to be modified.}

\item{track_apply}{An integer or vector of integers indicating which tracks (hh, kick, snare) to apply the modification to.}

\item{cc_parm}{If using cc modifier, this specifies the control change value.}

\item{cc_map}{If using cc modifier, this specifies the mapping of control changes.}
}
\value{
A modified MIDI sequence to be entered into the create_midi function.
}
\description{
Takes sequence produced by drum_machine function and applies modification to this sequence.
}
\examples{

# Applies double modification to first track (hh, or hi-hat) of drum sequence with .5 probability.
drum_machine(hh = 1:16, kick = seq(1, 16, by = 4), snare = c(5, 13)) \%>\%
random_modify(., modifier = "double", prob = .5, track_apply = 1)


}
