% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_muter.R
\name{random_muter}
\alias{random_muter}
\title{Randomly mute MIDI sequence}
\usage{
random_muter(seq_arg = NULL, prob = 1, position = "all",
  track_apply = NULL)
}
\arguments{
\item{seq_arg}{The sequence to be modified.  This should be the output of the drum_machine function.}

\item{prob}{(default = 1) The probability a note in the sequence will be muted.}

\item{position}{(default = "all") The position of items in the sequence to be muted.}

\item{track_apply}{An integer or vector of integers indicating which tracks (hh, kick, snare) to apply the mute to.}
}
\value{
A modified MIDI sequence to be entered into the create_midi function.
}
\description{
Takes sequence produced by drum_machine function and applies mutes to this sequence.
}
\examples{
# Applies mutes to first track (hh, or hi-hat) of drum sequence with .5 probability.

drum_machine(hh = 1:16, kick = seq(1, 16, by = 4), snare = c(5, 13)) \%>\%
random_muter(., prob = .5, track_apply = 1)


}
