% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiPrepareNotes.R
\name{toDeltaTime}
\alias{toDeltaTime}
\title{converts midi note DF to delta time based note DF}
\usage{
toDeltaTime(midiNotesDF)
}
\arguments{
\item{midiNotesDF}{DF of pitch velocity start end ticks}
}
\value{
DF that encodes pitch, velocity, delta time, and note-on and note-off and raw messages
}
\description{
Changes notes DF that encodes notes as pitch, velocity, start end tick to
DF that encodes pitch, velocity, delta time, and note-on and note-off.
Also adds raw midi bytes to the data frame for later generate of midi messages
}
