% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMidiFile.R
\name{writeMidiFile}
\alias{writeMidiFile}
\title{Write Midi File
#' Generate bytes for one midi Track}
\usage{
writeMidiFile(
  midiFileName,
  notesMidiDF,
  trackname = "Track",
  instrument = "Instrument"
)
}
\arguments{
\item{midiFileName}{the file name to write}

\item{notesMidiDF}{data frame of pitch, velocity, duration and beat}

\item{trackname}{name of track - character string}

\item{instrument}{name of instrument - character string}
}
\description{
write midiDF as midi rile
}
